requirements.txt
# Core libraries
numpy==1.26.4
pandas==2.2.2
opencv-python==4.9.0.80
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.5.0

# PyTorch (CPU version)
torch==2.3.0
torchvision==0.18.0

# GUI and drag-and-drop
tkinterdnd2==0.3.0
Pillow==10.3.0

# GUI backend support (ensure Tkinter is available in system)
# Note: `tkinter` is part of the Python standard library but needs to be installed with your Python distro.
